refactor: fix paths with unnecessary symbols
